apps/prompt/backend/pyproject.toml
[tool.poetry]
name = "prompt-manager-backend"
version = "0.1.0"
description = "Backend for Prompt Manager Tool"
authors = ["Rene Jean Corneille <rene-jean.corneille@onclusive.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.8.16"
core = {path = "../../../libs/core"}
openai = "1.3.4"
python-slugify = "8.0.1"
redis = "4.5.5"
python-redis-cache = "2.1.2"
PyGithub = "2.2.0"
dyntastic = "^0.14.0"
langchain = "^0.1.13"


[tool.poetry.group.dev.dependencies]
pytest = "7.2.0"
httpx = "0.23.3"
freezegun = "1.2.2"
fakeredis = {version = "2.14.1", extras = ["lua"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
